Moving mod_operation to arithmetics plugin
 -flyby: adapt includes for changes in HPX
