+++ phylanx/plugins/arithmetics/arithmetics.hpp
@@ -14,6 +14,7 @@
 #include <phylanx/plugins/arithmetics/generic_operation_bool.hpp>
 #include <phylanx/plugins/arithmetics/maximum.hpp>
 #include <phylanx/plugins/arithmetics/minimum.hpp>
+#include <phylanx/plugins/arithmetics/mod_operation.hpp>
 #include <phylanx/plugins/arithmetics/mul_operation.hpp>
 #include <phylanx/plugins/arithmetics/sub_operation.hpp>
 #include <phylanx/plugins/arithmetics/unary_minus_operation.hpp>
